+++ api/resources/routes.py
@@ -1,0 +1,9 @@
+from .requests import *
+
+def initialize_routes(api):
+    api.add_resource(Productos_lista, '/productos/<int:page>')
+    api.add_resource(Producto_registrar, '/productos/registrar/')
+    api.add_resource(Usuarios_lista, '/usuarios/<int:page>/')
+    api.add_resource(Usuarios_registrar, '/usuarios/registrar/')
+    api.add_resource(Usuario_detalle, '/usuario/<int:usuario_id>/')
+    api.add_resource(Login, '/login/')
