+++ makemigrations.sh
@@ -1,0 +1,3 @@
+cd api
+export FLASK_APP=app.py
+flask db migrate
