+++ start_api.sh
@@ -1,4 +1,4 @@
 cd api
-export FLASK_APP=project.py
+export FLASK_APP=app_api.py
 export FLASK_ENV=development
 python app_api.py
